+++ src/components/book/detail/BookItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-const BookDetailItem = ({ title }: { title: string }) => {
+const BookItem = ({ title }: { title: string }) => {
   return <article className='min-h-[8rem] border-basic'>{title}</article>;
 };
 
-export default BookDetailItem;
+export default BookItem;
