+++ src/components/book/detail/BookList.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 
 import { BooksProps, fetchDummy } from '@/pages/book/detail';
 
-import BookDetailItem from './BookDetailItem';
+import BookItem from './BookItem';
 
-const BookDetailList = () => {
+const BookList = () => {
   const { data: bookdetailList } = useQuery<BooksProps[]>(
     ['bookList'],
     fetchDummy,
@@ -14,10 +14,10 @@ const BookDetailList = () => {
   return (
     <section className='grid grid-cols-3 4 gap-1 overflow-scroll'>
       {bookdetailList?.map((item) => (
-        <BookDetailItem key={item.id} title={item.title} />
+        <BookItem key={item.id} title={item.title} />
       ))}
     </section>
   );
 };
 
-export default BookDetailList;
+export default BookList;
